notebooks/agent_webpage_helloworld/pyproject.toml
[tool.poetry]
name = "agent-webpage-helloworld"
version = "0.1.0"
description = ""
authors = ["Michael Deng <michaeldeng2003@gmail.com>"]
readme = "README.md"
package-mode = false


[tool.poetry.dependencies]
python = "==3.11.9"
openai = "^1.35.14"
python-dotenv = "^1.0.1"
jupyterlab = "^4.2.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
